[X] Spawn io_part onto dedicated executor
This now compiles and runs correctly, but interestingly enough, throws
timeout errors at exactly the same point (both in debug and release
builds) and it does when compiling without dedicated executor! That's
very interesting to me, there's a bunch of code different between the
two paths so the fact that we see identical performance is interesting.
I need to think on what that could be telling me more.